Reworked structure and updated docs.
This commit makes two major groups of changes:

1. It moves all specification content into a new `spec/` folder.
2. It updates the project's documentation files to clarify licensing,
   contribution guidelines, and more.

For the first change, this is really about clearly separating project
documentation from the specification itself. Previously, the project
was a mix of Markdown and text files organized in a single folder
together, which left it potentially unclear which files are part of
the specification, and which aren't.

The second set of changes includes a full README file, along with
formatting and grammatical changes to the other files.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>