First pass at a spec introduction.
This commit includes a first pass at writing an introduction for
the OmniBOR specification, centered around rough definitions of the
three key concepts for OmniBOR: Artifact Identifiers, Input Manifests,
and Artifact Dependency Graphs. It tries to introduce them in order,
and then to describe the benefits of having this information on hand.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>