feat: Add newline normalization to spec.
This updates the spec to reflect the need to do newline
normalization for cross-platform consistency purposes. Without
this normalization, files which are identical but are checked
out on Windows vs. Unix systems in Git will have different
identifiers, which is not ideal.

This normalization is already implemented in the OmniBOR Rust
implementation.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>